offload expect-elem-flag-behavior into test util for re-use on spinbox + simplify it to grab initialValue automatically
